test: split github module coverage
